Refactor log and errors package
