+++ observation-service/config/config.go
@@ -13,41 +13,13 @@ import (
 type Config struct {
 	Port int `envconfig:"PORT" default:"9001"`
 
-	DeploymentConfig  DeploymentConfig
+	DeploymentConfig  common_config.DeploymentConfig
 	NewRelicConfig    newrelic.Config
 	SentryConfig      sentry.Config
 	LogConsumerConfig LogConsumerConfig
 	LogProducerConfig LogProducerConfig
 	MonitoringConfig  MonitoringConfig
 }
-
-// DeploymentConfig captures the config related to the deployment of Observation Service
-type DeploymentConfig struct {
-	// EnvironmentType describes the environment Observation Service is deployed in
-	EnvironmentType string `default:"local"`
-	// ProjectName describes the CaraML project Observation Service is deployed for
-	ProjectName string `default:""`
-	// ServiceName describes the CaraML Observation Service name
-	ServiceName string `default:""`
-	// LogLevel captures the selected supported logging level
-	LogLevel LogLevel `split_words:"false" default:"INFO"`
-	// Maximum no. of go-routines that is allowed
-	MaxGoRoutines int `default:"1000"`
-}
-
-// LogLevel type is used to capture the supported logging levels
-type LogLevel string
-
-const (
-	// DebugLevel is used for verbose logs at debug level
-	DebugLevel LogLevel = "DEBUG"
-	// InfoLevel is used for logs that are info level and higher
-	InfoLevel LogLevel = "INFO"
-	// WarnLevel is used for logs that are warning level and higher
-	WarnLevel LogLevel = "WARN"
-	// ErrorLevel is used for logs that are error level and higher
-	ErrorLevel LogLevel = "ERROR"
-)
 
 // ObservationLoggerConsumerKind captures the consumer config for reading Observation Service logs
 type ObservationLoggerConsumerKind = string
