+++ observation-service/logger/kafka_consumer.go
@@ -12,8 +12,8 @@ import (
 	"github.com/confluentinc/confluent-kafka-go/kafka"
 	"google.golang.org/protobuf/proto"
 
+	"github.com/caraml-dev/timber/common/log"
 	"github.com/caraml-dev/timber/observation-service/config"
-	"github.com/caraml-dev/timber/observation-service/log"
 	"github.com/caraml-dev/timber/observation-service/monitoring"
 	"github.com/caraml-dev/timber/observation-service/services"
 	"github.com/caraml-dev/timber/observation-service/types"
