+++ observation-service/logger/kafka_producer.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/confluentinc/confluent-kafka-go/kafka"
 
+	"github.com/caraml-dev/timber/common/log"
 	"github.com/caraml-dev/timber/observation-service/config"
-	"github.com/caraml-dev/timber/observation-service/log"
 	"github.com/caraml-dev/timber/observation-service/monitoring"
 	"github.com/caraml-dev/timber/observation-service/services"
 	"github.com/caraml-dev/timber/observation-service/types"
