+++ observation-service/logger/observation_logger.go
@@ -6,8 +6,8 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/caraml-dev/timber/common/log"
 	"github.com/caraml-dev/timber/observation-service/config"
-	"github.com/caraml-dev/timber/observation-service/log"
 	"github.com/caraml-dev/timber/observation-service/monitoring"
 	"github.com/caraml-dev/timber/observation-service/services"
 	"github.com/caraml-dev/timber/observation-service/types"
