+++ observation-service/server/server.go
@@ -17,11 +17,11 @@ import (
 	"google.golang.org/grpc/health/grpc_health_v1"
 	"google.golang.org/grpc/reflection"
 
+	customErr "github.com/caraml-dev/timber/common/errors"
+	"github.com/caraml-dev/timber/common/log"
 	"github.com/caraml-dev/timber/observation-service/appcontext"
 	"github.com/caraml-dev/timber/observation-service/config"
 	"github.com/caraml-dev/timber/observation-service/controller"
-	customErr "github.com/caraml-dev/timber/observation-service/errors"
-	"github.com/caraml-dev/timber/observation-service/log"
 )
 
 var (
