+++ observation-service/services/metric_service.go
@@ -7,8 +7,9 @@ import (
 
 	"github.com/gojek/mlp/api/pkg/instrumentation/metrics"
 
+	common_config "github.com/caraml-dev/timber/common/config"
+	"github.com/caraml-dev/timber/common/log"
 	"github.com/caraml-dev/timber/observation-service/config"
-	"github.com/caraml-dev/timber/observation-service/log"
 	"github.com/caraml-dev/timber/observation-service/monitoring"
 )
 
@@ -22,11 +23,11 @@ type MetricService interface {
 
 type metricService struct {
 	Kind             config.MetricSinkKind
-	DeploymentConfig config.DeploymentConfig
+	DeploymentConfig common_config.DeploymentConfig
 }
 
 // NewMetricService initializes a metricService struct
-func NewMetricService(deploymentCfg config.DeploymentConfig, monitoringCfg config.MonitoringConfig) (MetricService, error) {
+func NewMetricService(deploymentCfg common_config.DeploymentConfig, monitoringCfg config.MonitoringConfig) (MetricService, error) {
 	switch monitoringCfg.Kind {
 	case config.NoopMetricSink:
 	case config.PrometheusMetricSink:
