+++ observation-service/services/metric_service_test.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/stretchr/testify/suite"
 
+	common_config "github.com/caraml-dev/timber/common/config"
 	"github.com/caraml-dev/timber/observation-service/config"
 	"github.com/caraml-dev/timber/observation-service/internal/testutils"
 	"github.com/caraml-dev/timber/observation-service/monitoring"
@@ -23,7 +24,7 @@ func (s *MetricServiceTestSuite) SetupSuite() {
 
 	var err error
 	s.cfg = config.Config{
-		DeploymentConfig: config.DeploymentConfig{
+		DeploymentConfig: common_config.DeploymentConfig{
 			ProjectName: "default",
 			ServiceName: "observation-service",
 		},
